refactor: move resolvers and queries to common file
