--- src/routes/graphql/queries/memberTypesQuery.ts
@@ -1,8 +1,0 @@
-import { GraphQLList } from 'graphql';
-import { memberTypesResolver } from '../resolvers/memberTypesResolver';
-import { memberTypeType } from '../types/memberTypeType';
-
-export const memberTypesQuery = {
-  type: new GraphQLList(memberTypeType),
-  resolve: memberTypesResolver,
-};
