--- src/routes/graphql/resolvers/memberTypesResolver.ts
@@ -1,9 +1,0 @@
-import { FastifyInstance } from 'fastify';
-
-export async function memberTypesResolver(
-  _parent: any,
-  _args: any,
-  context: FastifyInstance
-) {
-  return await context.db.memberTypes.findMany();
-}
