--- src/routes/graphql/resolvers/postsResolver.ts
@@ -1,9 +1,0 @@
-import { FastifyInstance } from 'fastify';
-
-export async function postsResolver(
-  _parent: any,
-  _args: any,
-  context: FastifyInstance
-) {
-  return await context.db.posts.findMany();
-}
